Various fixes and additions...
Added a heal verb for admins. Explanation for the lack of a verb
for developers. A category for the shared special verbs. Moved
mob procedures to their own file. A heal_all (completely heals you)
proc for health. A proper needs system, which begins incrementing
at creation of a mob (may need balancing later), needs to the statpanel
in a vague way-- Health needs to be changed like this later. Made attack
verb only work on mobs and inform you of the name. A few small edits to
turfs, and debugging map.