A few changes, additions, and reworking of AI...
Added a message to the top of mobs in order to explain what to do
if you want to create a new mob.

Rewrote AIs entirely into datums. Simply set the default variable of
any mob to 1 in order to turn it into an AI. Currently has a structure
of checking for behaviors, updating the behavior with the result, and
then performing actions based upon the behavior. Needs a fair bit of work,
but it looks like a solid frame to work with.

... had to set myself as the only developer via the developer verb for
the time being. I don't like it, but it does mean that I can update the
game if I choose to host it, and people can still look around/have it
open-source. Will remove/change in the future if the project manages to
go anywhere.

Changed a message when you fail to attack, and added a check to make sure
that the target happened to be an NPC before trying to call the hit_by
proc, so as to avoid runtimes.

Made the map slightly funner! :D